微信开放平台代码分离

+++ koala-common/koala-wechat-sdk/src/main/java/org/igetwell/wechat/sdk/bean/card/qrcode/QrCodeCreate.java
@@ -11,7 +11,7 @@ public class QrCodeCreate {
     /**
      * 指定二维码的有效时间，范围是60 ~ 1800秒。不填默认为365天有效
      */
-    private int expireSeconds;
+    private Long expireSeconds;
 
     /**
      * 卡券领取类型：QR_CARD
